requirements.txt
flask
flask-cors
python-dotenv
openai
llama-index
gunicorn
